+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.3.1
 ----------------
 - Fix wkhtmltopdf error on Windows: prepend file:/// in PDF.php
+- Fix PHP error removed s.*, select each student field in Export.php
 
 Changes in 3.3
 --------------